read session history from disk in main process, remove sidecar round-trip
Move history reading from sidecar IPC to a direct disk read in the
main process via SessionManager.continueRecent(). Simpler, faster,
and doesn't depend on the sidecar being ready.
